requirements.txt
alembic==1.9.1
anyio==3.6.2
apache-airflow==2.5.0
apache-airflow-backport-providers-amazon==2021.3.3
apache-airflow-providers-apache-hdfs==3.2.0
apache-airflow-providers-apache-hive==5.1.1
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-common-sql==1.3.2
apache-airflow-providers-ftp==3.3.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-slack==7.2.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==1.12.3
attrs==22.2.0
Babel==2.11.0
blinker==1.5
boto3==1.15.18
botocore==1.18.18
cached-property==1.5.2
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.0.2
commonmark==0.9.1
configparser==3.5.3
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.32
croniter==0.3.37
cryptography==39.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.1.0
fastavro==1.7.0
Flask==2.2.2
Flask-Admin==1.5.4
Flask-AppBuilder==4.1.4
Flask-Babel==1.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==1.1.0
funcsigs==1.0.2
future==0.18.3
graphviz==0.20.1
greenlet==2.0.1
gssapi==1.8.2
gunicorn==20.1.0
h11==0.14.0
hdfs==2.7.0
hmsclient==0.1.1
httpcore==0.16.3
httpx==0.23.3
idna==2.10
importlib-metadata==2.1.3
importlib-resources==5.10.2
inflection==0.5.1
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
krb5==0.4.1
lazy-object-proxy==1.4.3
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
natsort==8.2.0
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pathspec==0.9.0
pendulum==2.1.2
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
prison==0.2.1
protobuf==3.20.3
psutil==5.9.4
pure-sasl==0.6.2
py4j==0.10.9.5
pycparser==2.21
Pygments==2.14.0
PyHive==0.6.5
PyJWT==2.6.0
pyrsistent==0.19.3
pyspark==3.3.1
pyspnego==0.7.0
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-kerberos==0.14.0
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.0.1
s3transfer==0.3.7
sasl==0.3.1
setproctitle==1.3.2
six==1.16.0
slack-sdk==3.19.5
snakebite-py3==3.0.5
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.1.0
termcolor==2.2.0
text-unidecode==1.3
thrift==0.16.0
thrift-sasl==0.4.3
typing-extensions==4.4.0
tzlocal==1.5.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.25.11
watchtower==0.7.3
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.11.0
zope.deprecation==4.4.0
